Added state-based completing of tasks to updated frontend layout.
Function & endpoint added to support getting all subdomains & tasks for those subdomains to the backend. This endpoint will return all subdomains with a task attribute that contains an array of tasks, which are grouped by task group and ordered by their order within the group. Adding, updating, or removing tasks using the backend endpoints will maintain correct ordering. By default, tasks are added to a new group (if no group or order is specified), added as last in order (if group is specified), or at the correct group and order (other tasks are reordered accordingly, and a valid order number is assumed).