Implemented React Routing, User Auth & Login.
React Routing has been implemented, and there is now a second page (Login page) that is automatically routed to if a token isn't found in in localStorage. Passwords are hashed & salted, and all endpoints save for user creation authenticate against a saved token. Token expires after 1 day. If localStorage is cleared, on next load or refresh, user will be kicked back to the login page and locked out.